+++ bouncer/shared/evm_vault_swap.ts
@@ -22,7 +22,7 @@ import {
   newAddress,
 } from './utils';
 import { getBalance } from './get_balance';
-import { CcmDepositMetadata, DcaParams, FillOrKillParamsX128 } from '../shared/new_swap';
+import { CcmDepositMetadata, DcaParams, FillOrKillParamsX128 } from './new_swap';
 import { SwapContext, SwapStatus } from './swap_context';
 
 const erc20Assets: Asset[] = ['Flip', 'Usdc', 'Usdt', 'ArbUsdc'];
