+++ bouncer/shared/swapping.ts
@@ -16,7 +16,7 @@ import {
 } from '../shared/utils';
 import { BtcAddressType } from '../shared/new_btc_address';
 import { CcmDepositMetadata } from '../shared/new_swap';
-import { performVaultSwap } from '../shared/vault_swap';
+import { performVaultSwap } from './evm_vault_swap';
 import { SwapContext, SwapStatus } from './swap_context';
 
 enum SolidityType {
