+++ bouncer/tests/btc_vault_swap.ts
@@ -14,15 +14,13 @@ import { getChainflipApi, observeEvent } from '../shared/utils/substrate';
 /* eslint-disable @typescript-eslint/no-use-before-define */
 export const testBtcVaultSwap = new ExecutableTest('Btc-Vault-Swap', main, 60);
 
-interface EncodedSwapRequest {
-  Bitcoin: {
-    nulldata_utxo: string;
-  };
-}
-
 interface VaultSwapDetails {
-  deposit_address: string;
-  encoded_params: EncodedSwapRequest;
+  chain: {
+    Bitcoin: {
+      nulldata_utxo: string;
+      deposit_address: string;
+    };
+  };
 }
 
 async function buildAndSendBtcVaultSwap(
@@ -40,25 +38,22 @@ async function buildAndSendBtcVaultSwap(
   const feeBtc = 0.00001;
   const { inputs, change } = await selectInputs(Number(depositAmountBtc) + feeBtc);
 
-  const vaultSwapDetails = (await chainflip.rpc(
-    `cf_get_vault_swap_details`,
-    broker.address,
-    'BTC', // source_asset
-    destinationAsset.toUpperCase(),
-    { [shortChainFromAsset(destinationAsset)]: destinationAddress },
-    0, // broker_commission
-    0, // min_output_amount
-    0, // retry_duration
-  )) as unknown as VaultSwapDetails;
-  testBtcVaultSwap.debugLog(
-    'nulldata_utxo:',
-    vaultSwapDetails.encoded_params.Bitcoin.nulldata_utxo,
-  );
+  const vaultSwapDetails = (
+    (await chainflip.rpc(
+      `cf_get_vault_swap_details`,
+      broker.address,
+      'BTC', // source_asset
+      destinationAsset.toUpperCase(),
+      { [shortChainFromAsset(destinationAsset)]: destinationAddress },
+      0, // broker_commission
+      0, // min_output_amount
+      0, // retry_duration
+    )) as unknown as VaultSwapDetails
+  ).chain.Bitcoin;
+  testBtcVaultSwap.debugLog('nulldata_utxo:', vaultSwapDetails.nulldata_utxo);
 
   // The `createRawTransaction` function will add the op codes, so we have to remove them here.
-  const nullDataWithoutOpCodes = vaultSwapDetails.encoded_params.Bitcoin.nulldata_utxo
-    .replace('0x', '')
-    .substring(4);
+  const nullDataWithoutOpCodes = vaultSwapDetails.nulldata_utxo.replace('0x', '').substring(4);
 
   const outputs = [
     {
